+++ packages/client/src/typed-codecs/shaped-codec.ts
@@ -1,9 +1,4 @@
-import {
-  Binary,
-  Codec,
-  FixedSizeBinary,
-  ResultPayload,
-} from "@polkadot-api/substrate-bindings"
+import { Binary, Codec, ResultPayload } from "@polkadot-api/substrate-bindings"
 
 type Tuple<T> = [T, ...T[]]
 type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (
@@ -36,7 +31,6 @@ type _Inner<T> = T extends void
           | symbol
           | Uint8Array
           | Binary
-          | FixedSizeBinary<any>
       ? {}
       : T extends Tuple<any>
         ? {