+++ packages/merkleize-metadata/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-## 1.1.22 to 1.1.27 - 2025-10-31
+### Fixed
+
+- Find first-class signed extensions only in `ExtVersion` 0
+
+## 1.1.22 to 1.1.28 - 2026-01-05
 
 ### Fixed