+++ packages/observable-client/CHANGELOG.md
@@ -12,6 +12,12 @@
 
 ## Unreleased
 
+## 0.17.2 - 2026-01-05
+
+### Fixed
+
+- Update dependencies
+
 ## 0.17.1 - 2025-12-08
 
 ### Fixed