+++ packages/signers/signer/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-## 0.2.1 to 0.2.11 - 2025-10-31
+### Fixed
+
+- Use right signed extensions for transaction building
+
+## 0.2.1 to 0.2.12 - 2026-01-05
 
 ### Fixed